From Exploration to Structure (#4)
On the surface, this release doesn't change much or provide drastically new functionality, but it begins to lay the foundation for something that could evolve in the long-term.

- Mostly refactoring to support the long-term vision
- Add Reek to dev dependencies
- Enable Inch in the default commands
- Reduce external dependencies
- Broaden official support Ruby 2.5.9, 2.6.8, 2.7.4, and 3.0.2
- Add more robust GitHub Actions integration
- Add Code Coverage via SimpleCov and set the bar at 100%
- Begin to expand documentation coverage
- Improve UX of results, timing, output, and error recovery guidance

+++ lib/reviewer/history.rb
@@ -26,7 +26,13 @@ module Reviewer
     end
 
     def reset!
+      return unless File.exist?(file)
+
       FileUtils.rm(file)
+    end
+
+    def self.reset!
+      new.reset!
     end
   end
 end
